refactor: Turip 관련 도메인 패키지 위치 수정 (folder -> turip)
