+++ android/app/src/main/java/com/on/turip/data/turip/TuripMapper.kt
@@ -9,8 +9,8 @@ import com.on.turip.data.turip.dto.TuripResponse
 import com.on.turip.data.turip.dto.TuripsByPlaceResponse
 import com.on.turip.data.turip.dto.TuripsResponse
 import com.on.turip.domain.bookmark.TuripPlace
-import com.on.turip.domain.folder.Turip
 import com.on.turip.domain.trip.Place
+import com.on.turip.domain.turip.Turip
 
 fun TuripsResponse.toDomain(): List<Turip> = turipsResponse.map { it.toDomain() }
