+++ android/app/src/main/java/com/on/turip/data/turip/repository/DefaultTuripRepository.kt
@@ -8,8 +8,8 @@ import com.on.turip.data.turip.toDomain
 import com.on.turip.data.turip.toPatchRequestDto
 import com.on.turip.data.turip.toPostRequestDto
 import com.on.turip.domain.bookmark.TuripPlace
-import com.on.turip.domain.folder.Turip
-import com.on.turip.domain.folder.repository.TuripRepository
+import com.on.turip.domain.turip.Turip
+import com.on.turip.domain.turip.repository.TuripRepository
 import javax.inject.Inject
 
 class DefaultTuripRepository @Inject constructor(
