+++ android/app/src/main/java/com/on/turip/domain/turip/TuripNameStatus.kt
@@ -1,4 +1,4 @@
-package com.on.turip.domain.folder
+package com.on.turip.domain.turip
 
 enum class TuripNameStatus {
     OK,
